Something is Something
It has the CRUD of Shop, providers and products. Each shop has a list of providers and products. When you create a shop and access to its details, you have the possibility of accesing to the CRUD of its products and providers. After that,  you can't come back to the shop detail in this version.